feat: warn editor when review URL is missing/wrong in reviewer notification
