+++ src/templates/admin/elements/email_form.html
@@ -1,12 +1,12 @@
 {% load foundation %}
-<form method="POST" enctype="multipart/form-data">
+<form method="POST" enctype="multipart/form-data"{% if form_id %} id="{{ form_id }}"{% endif %}>
     {% csrf_token %}
     <div class="card-section">
         {{ form | foundation }}
     </div>
     <div class="card-divider">
         <div class="button-group">
-            <button type="submit" class="button success" name="send"><i class="fa fa-envelope-o">&nbsp;</i>Send</button>
+            <button type="submit" class="button success" name="{{ button_name|default:'send' }}"><i class="fa fa-envelope-o">&nbsp;</i>Send</button>
             {% if skip %}
             <button type="submit" class="button warning" name="skip" value="skip"><i class="fa fa-step-forward">&nbsp;</i>Skip & continue</button>
             {% endif %}
